+++ packages/core/src/StateMachine.ts
@@ -144,6 +144,7 @@ export class StateMachine<
 
     this.transition = this.transition.bind(this);
     this.getInitialSnapshot = this.getInitialSnapshot.bind(this);
+    this.getPersistedSnapshot = this.getPersistedSnapshot.bind(this);
     this.restoreSnapshot = this.restoreSnapshot.bind(this);
     this.start = this.start.bind(this);